+++ src/Contracts/Tracker.php
@@ -1,4 +1,7 @@
 <?php namespace Arcanedev\LaravelTracker\Contracts;
+use Exception;
+use Illuminate\Http\Request;
+use Illuminate\Routing\Route;
 
 /**
  * Interface  Tracker
@@ -12,6 +15,28 @@ interface Tracker
      |  Main Functions
      | ------------------------------------------------------------------------------------------------
      */
+    /**
+     * Start the tracking.
+     *
+     * @param  \Illuminate\Http\Request $request
+     */
+    public function trackRequest(Request $request);
+
+    /**
+     * Track the matched route.
+     *
+     * @param  \Illuminate\Routing\Route  $route
+     * @param  \Illuminate\Http\Request   $request
+     */
+    public function trackMatchedRoute(Route $route, Request $request);
+
+    /**
+     * Track the exception.
+     *
+     * @param  \Exception  $exception
+     */
+    public function trackException(Exception $exception);
+
     /**
      * Enable the tracking.
      */
